Packets: Replaced 16-bit unsigned integer with 64-bit
16-bit unsigned integers doesn't save that much more space to warrant
using over the 64-bit one

On my machine, we're only saving around 6 bytes using 16-bit

+++ include/BaconEngine/BaconEngine/Server/Packet.h
@@ -29,8 +29,8 @@ typedef void (*BE_Packet_Run)(BE_Client client, char data[BE_PACKET_MAXIMUM_DATA
 
 // TODO: Flags
 
-BE_BINARYEXPORT void BE_Packet_Register(uint16_t operationCode, BA_Boolean acceptUnconnected, BE_Packet_Run Run);
+BE_BINARYEXPORT void BE_Packet_Register(uint64_t operationCode, BA_Boolean acceptUnconnected, BE_Packet_Run Run);
 
-BE_BINARYEXPORT void BE_Packet_Send(BE_Client client, uint16_t operationCode, char data[BE_PACKET_MAXIMUM_DATA]);
+BE_BINARYEXPORT void BE_Packet_Send(BE_Client client, uint64_t operationCode, char data[BE_PACKET_MAXIMUM_DATA]);
 BA_CPLUSPLUS_SUPPORT_GUARD_END()
